更新实体类 添加商铺插入测试

+++ src/main/java/com/li/entity/Area.java
@@ -3,18 +3,18 @@ package com.li.entity;
 import java.util.Date;
 
 public class Area {
-    private Long areaId;
+    private Integer areaId;
     private String areaName;
     private String areaDesc;
     private Integer priority;
     private Date createTime;
     private Date lastEditTime;
 
-    public Long getAreaId() {
+    public Integer getAreaId() {
         return areaId;
     }
 
-    public void setAreaId(Long areaId) {
+    public void setAreaId(Integer areaId) {
         this.areaId = areaId;
     }
